moved components to route folder
