+++ app/shopping-list/components/IngredientNode.tsx
