+++ app/shopping-list/components/NewIngredientBar.tsx
